requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
newspaper3k==0.2.8
requests==2.31.0

# AI/ML
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
faiss-cpu==1.7.4
numpy==1.24.3

# LLM APIs
openai==1.3.0
groq==0.4.1
google-generativeai==0.3.0

# UI
streamlit==1.28.0
streamlit-chat==0.1.1
plotly==5.15.0
pandas==2.0.3

# Utilities
aiohttp==3.9.0
asyncio-throttle==1.0.2
python-dateutil==2.8.2